Updates for latest version of Essentials

+++ CustomCarouselPath/ViewController.m
@@ -95,7 +95,7 @@
 
 #pragma mark - SEssentialsCarouselDataSource methods
 
--(int)numberOfItemsInCarousel:(SEssentialsCarousel *)carousel
+-(NSUInteger)numberOfItemsInCarousel:(SEssentialsCarousel *)carousel
 {
     return _numberOfItemsInCarousel;
 }
